Move payload coauthor handling into a dedicated service
Injecting the coauthor context plus a trait pushed PatchService to 21
class dependencies (php:S1200, max 20). An Element service replaces both,
so each caller depends on one collaborator instead of two and the class
drops back under the limit.
